9/13 passed tests for task 6 Identified a problem with our caclulateFreeSeats() so edited it into something that I think will work better for us in our future work.

+++ lab4/c/lab4c_schema.sql
@@ -113,7 +113,7 @@ create table reservation(
 
 -- Create table credit_card 
 create table credit_card(
-       card_nr integer not null,
+       card_nr bigint not null,
        holder varchar(60) not null, -- First name + last name = 30 + 30 
 
        constraint pk_credit_card primary key(card_nr));
@@ -129,7 +129,7 @@ create table booked_pass(
 -- Create table booked 
 create table booked(
        reservation integer not null,
-       card integer not null,
+       card bigint not null,
        total_price double not null,       
        
        constraint pk_booked primary key(reservation));
